Move sales-terminal to new closer component
